requirements.txt
fastapi
python-multipart
uvicorn[standard]
websockets
deepgram-sdk
openai
pyngrok
python-dotenv
requests
